Refactor Employee class: add task creation and progress retrieval methods; remove Leader class
